extract properties to yaml conversion into a separate class
